requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
Pillow==8.1.1
PyJWT==1.7.1
pypi==2.1
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.25.9
whitenoise==5.2.0
dj-database-url==0.5.0
django-storages==1.8
boto3==1.9.234
django-heroku==0.3.1
psycopg2
